- Began refactoring FlexBuilder integration into a separate feature project  - Added extension point to provide builder integration for IDE bridge projects  - Moved some of the processors into the right package

+++ plugin/trunk/src/com/dtsworkshop/flextools/codemodel/WorkingSpaceModelStateManager.java
@@ -99,6 +99,10 @@ public class WorkingSpaceModelStateManager extends AbstractStateManager implemen
 
 	public boolean removeProjectState(IProject project) {
 		ProjectStateEntry entry = getEntry(project);
+		if(entry == null) {
+			// Somehow we've not got the entry for the project, so we'll just report back that we succeeded in removing it...
+			return true;
+		}
 		File projectDirectory = entry.getWorkingSpace().toFile();
 		boolean success = projectDirectory.delete();
 		if(success) {
